+++ udata_ods/__init__.py
@@ -4,5 +4,5 @@ OpenDataSoft integration for udata
 '''
 from __future__ import unicode_literals
 
-__version__ = '1.0.2.dev'
+__version__ = '1.1.0'
 __description__ = 'OpenDataSoft integration for udata'