+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Do not export SHP when above records limit [#26](https://github.com/opendatateam/udata-ods/pull/20)
 - Initial resources preview support [#29](https://github.com/opendatateam/udata-ods/pull/29) [#34](https://github.com/opendatateam/udata-ods/pull/34)
 - Use `udata.frontend.markdown:parse_html()` to parse description [#35](https://github.com/opendatateam/udata-ods/pull/35)
+- Attachments support [#36](https://github.com/opendatateam/udata-ods/pull/36)
 
 ## 1.0.1 (2018-03-13)